Refactoring utils to new MVP modules.
