+++ app/src/main/java/com/robotandpencils/app/util/PerFragmentUtil.java
@@ -2,7 +2,7 @@ package com.robotandpencils.app.util;
 
 import android.app.Fragment;
 
-import com.robotandpencils.app.ui.common.BaseFragmentModule;
+import com.robotandpencils.app.ui.common.view.BaseFragmentModule;
 
 import javax.inject.Inject;
 import javax.inject.Named;
